pyproject.toml
[tool.poetry]
name = "chatmate"
version = "0.1.0"
description = "Project to access weekly transactions as they appear in my gmail inbox"
authors = ["Leila Wolfe <leilanoellewolfe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10" # Assuming you are using Python 3.10 or newer

# 1. GMAIL API ACCESS (Essential for initial run)
google-api-python-client = "^2.187.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"

# 2. CORE DATA PROCESSING
requests = "^2.31.0"
#pandas = "^2.2.0"
#numpy = "^1.26.0"

# 3. EMAIL/HTML PARSING & REPORTING
# Replaced 'beautifulsoup' and 'bs4' with the standard, modern package: beautifulsoup4
beautifulsoup4 = "^4.12.3"
# Highly recommended, fast parser for BeautifulSoup
lxml = "^5.2.2"
jinja2 = "^3.1.4"
reportlab = "^4.1.0"
pydantic = "^2.7.0"

# 4. REMOVED FOR STABILITY: 
# Removed the following massive and conflict-prone packages 
# (seaborn, matplotlib, plotly, dash, google-genai, nltk, spacy, langchain)
# We can add these back in small, controlled groups later after confirming email access works.

[tool.poetry.group.dev.dependencies]
# 5. DEVELOPMENT TOOLS (Moved from main dependencies)
# These tools are only needed by developers (you) and are kept separate 
# to avoid conflicts with production dependencies.
pytest = "^8.2.0"
black = "^24.4.2"
ruff = "^0.14.5"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"